Refactoring OutputCtrl to pull logic into service
Created a summary service to handle calculating sums and totals,
You should try and keep your controllers as thin as possible and act
as glue between the model and the view.

Keeping with the idea of an event-driven approach, I moved the
event handlers for expense-added and income-added into the summary
service.

This is now responsible for calculating the totals, and then letting
the rest of the system know 'hey, totals are updated' - so anyone
that wants to respond to the updated totals is able to do so.

I generally try to keep my events as a 'X was done', and not as a
'please do something'.