Refactor OutputCtrl to use controllerAs
To keep things consistent with the other controllers that were
updated to controllerAs, just a quick refactor to get this one
in line with the others.

While we are trying to reduce the reliance on $scope, since we are
using events - we need to know when to unbind the event handlers.

When you get into more complex apps with routing/navigation and
controllers get created/destroyed as the user navigates around the
system we need to know when the controller is getting destroyed so
we can unbind the event handlers.